airflow/requirements.txt
psycopg2-binary==2.9.9
clickhouse-driver==0.2.6
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
